requirements.txt
Flask==3.0.0
asyncpg==0.29.0
python-dotenv==1.0.0
gunicorn==21.2.0
